[WMS-22] Move /BaseWarehouseDialogController to base package.
